requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==1.9.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
idna==2.8
mysqlclient==1.3.13
Pillow==5.4.1
PyMySQL==0.9.3
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
